Downloads/tech-trail-main/tech-trail-main/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
email-validator==2.2.0
passlib[bcrypt]==1.7.4
